refactor(core): port core plugin execution to TypeScript
- Rewrites the main plugin execution logic from JavaScript to TypeScript.
- Enhances code quality and maintainability through static type checking.
- Migrates module loading to ECMAScript Modules (ESM) using import.meta.url.
- Updates plugin loading to use dynamic import() for improved modularity.
- Ensures the external behavior and output of the plugin runner remain identical.
